Graph + Graphalloc OK
